blog_automation/requirements.txt
feedparser==6.0.11
PyPDF2==3.0.1
requests==2.32.3
PyYAML==6.0.2
openai==1.61.0
